Use of double instead of float for all Particles data.

+++ include/io/types.hpp
@@ -9,16 +9,16 @@
 namespace io {
 	namespace types {
 		struct ParticuleData {
-			float Pos[3];
-			float Vit[3];
-			float m;
-			float Pot;
-			float Acc[3];
-			float dAdt;
-			float ts;
-			float Rho;
-			float U;
-			float Ne;
+			double Pos[3];
+			double Vit[3];
+			double m;
+			double Pot;
+			double Acc[3];
+			double dAdt;
+			double ts;
+			double Rho;
+			double U;
+			double Ne;
 			int Id;
 			int Type;
 		};
